+++ lib/sidekiq/transaction_guard/middleware.rb
@@ -10,9 +10,7 @@ module Sidekiq
     # the default behavior set in `Sidekiq::TransactionGuard.mode` and
     # `Sidekiq::TransactionGuard.notify` respectively.
     class Middleware
-      if Gem::Version.new(Sidekiq::VERSION) >= Gem::Version.new("7.0")
-        include Sidekiq::ClientMiddleware
-      end
+      include Sidekiq::ClientMiddleware if defined?(Sidekiq::ClientMiddleware)
 
       def call(worker_class, job, queue, redis_pool)
         # Check if we need to log this. Also, convert worker_class to its actual class
@@ -68,7 +66,9 @@ module Sidekiq
         mode = worker_mode(job)
         return if mode == :disabled
 
-        message = "#{worker_class.name} was called from inside a database transaction"
+        message = "#{worker_class.name} was called from inside a database transaction. " \
+          "Resolve by moving the job outside the transaction, using an after_commit callback, " \
+          "or setting `sidekiq_options transaction_guard: :disabled` if the job is safe to run before the transaction commits."
         if mode == :error
           raise Sidekiq::TransactionGuard::InsideTransactionError.new(message)
         end