+++ lib/sidekiq/transaction_guard/railtie.rb
@@ -3,13 +3,8 @@
 module Sidekiq::TransactionGuard
   class Railtie < ::Rails::Railtie
     initializer "sidekiq.transaction_guard" do
-      Sidekiq::TransactionGuard.mode = ((Rails.env.development? || Rails.env.test?) ? :error : :warn)
-
-      Sidekiq.configure_client do |config|
-        config.client_middleware do |chain|
-          chain.add Sidekiq::TransactionGuard::Middleware
-        end
-      end
+      mode = (Rails.env.development? || Rails.env.test?) ? :error : :warn
+      Sidekiq::TransactionGuard.init(mode: mode)
     end
   end
 end